change profiles import that failed on extended clients
